refactor: move unmarshaling code to internal/codec
