+++ src/main/java/io/weaviate/client6/v1/collections/aggregate/WeaviateAggregate.java
@@ -1,13 +1,10 @@
 package io.weaviate.client6.v1.collections.aggregate;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.function.Consumer;
 
 import io.weaviate.client6.internal.GrpcClient;
 import io.weaviate.client6.internal.codec.grpc.v1.AggregateMarshaler;
+import io.weaviate.client6.internal.codec.grpc.v1.AggregateUnmarshaler;
 import io.weaviate.client6.v1.query.NearVector;
 
 public class WeaviateAggregate {
@@ -25,94 +22,20 @@ public class WeaviateAggregate {
         .addAggregation(aggregation)
         .marshal();
     var reply = grpcClient.grpc.aggregate(req);
-
-    Long totalCount = null;
-    Map<String, Metric.Values> properties = new HashMap<>();
-
-    if (reply.hasSingleResult()) {
-      var single = reply.getSingleResult();
-      totalCount = single.hasObjectsCount() ? single.getObjectsCount() : null;
-      var aggregations = single.getAggregations().getAggregationsList();
-      for (var agg : aggregations) {
-        var property = agg.getProperty();
-        Metric.Values value = null;
-
-        if (agg.hasInt()) {
-          var metrics = agg.getInt();
-          value = new IntegerMetric.Values(
-              metrics.hasCount() ? metrics.getCount() : null,
-              metrics.hasMinimum() ? metrics.getMinimum() : null,
-              metrics.hasMaximum() ? metrics.getMaximum() : null,
-              metrics.hasMean() ? metrics.getMean() : null,
-              metrics.hasMedian() ? metrics.getMedian() : null,
-              metrics.hasMode() ? metrics.getMode() : null,
-              metrics.hasSum() ? metrics.getSum() : null);
-        } else {
-          assert false : "branch not covered";
-        }
-        if (value != null) {
-          properties.put(property, value);
-        }
-      }
-    }
-    var result = new AggregateResponse(properties, totalCount);
-    return result;
+    return new AggregateUnmarshaler(reply).single();
   }
 
-  public AggregateGroupByResponse overAll(Consumer<AggregateGroupByRequest.GroupBy.Builder> groupByOptions,
+  public AggregateGroupByResponse overAll(
+      AggregateGroupByRequest.GroupBy groupBy,
       Consumer<AggregateRequest.Builder> options) {
     var aggregation = AggregateRequest.with(collectionName, options);
-    var groupBy = AggregateGroupByRequest.GroupBy.with(groupByOptions);
 
     var req = new AggregateMarshaler(aggregation.collectionName())
         .addAggregation(aggregation)
         .addGroupBy(groupBy)
         .marshal();
     var reply = grpcClient.grpc.aggregate(req);
-
-    List<Group<?>> groups = new ArrayList<>();
-    if (reply.hasGroupedResults()) {
-      for (var result : reply.getGroupedResults().getGroupsList()) {
-        final Long totalCount = result.hasObjectsCount() ? result.getObjectsCount() : null;
-
-        GroupedBy<?> groupedBy = null;
-        var gb = result.getGroupedBy();
-        if (gb.hasInt()) {
-          groupedBy = new GroupedBy<Long>(gb.getPathList().get(0), gb.getInt());
-        } else if (gb.hasText()) {
-          groupedBy = new GroupedBy<String>(gb.getPathList().get(0), gb.getText());
-        } else {
-          assert false : "branch not covered";
-        }
-
-        Map<String, Metric.Values> properties = new HashMap<>();
-        for (var agg : result.getAggregations().getAggregationsList()) {
-          var property = agg.getProperty();
-          Metric.Values value = null;
-
-          if (agg.hasInt()) {
-            var metrics = agg.getInt();
-            value = new IntegerMetric.Values(
-                metrics.hasCount() ? metrics.getCount() : null,
-                metrics.hasMinimum() ? metrics.getMinimum() : null,
-                metrics.hasMaximum() ? metrics.getMaximum() : null,
-                metrics.hasMean() ? metrics.getMean() : null,
-                metrics.hasMedian() ? metrics.getMedian() : null,
-                metrics.hasMode() ? metrics.getMode() : null,
-                metrics.hasSum() ? metrics.getSum() : null);
-          } else {
-            assert false : "branch not covered";
-          }
-          if (value != null) {
-            properties.put(property, value);
-          }
-        }
-        Group<?> group = new Group<>(groupedBy, properties, totalCount);
-        groups.add(group);
-
-      }
-    }
-    return new AggregateGroupByResponse(groups);
+    return new AggregateUnmarshaler(reply).grouped();
   }
 
   public AggregateResponse nearVector(
@@ -127,48 +50,16 @@ public class WeaviateAggregate {
         .addNearVector(nearVector)
         .marshal();
     var reply = grpcClient.grpc.aggregate(req);
-
-    Long totalCount = null;
-    Map<String, Metric.Values> properties = new HashMap<>();
-
-    if (reply.hasSingleResult()) {
-      var single = reply.getSingleResult();
-      totalCount = single.hasObjectsCount() ? single.getObjectsCount() : null;
-      var aggregations = single.getAggregations().getAggregationsList();
-      for (var agg : aggregations) {
-        var property = agg.getProperty();
-        Metric.Values value = null;
-
-        if (agg.hasInt()) {
-          var metrics = agg.getInt();
-          value = new IntegerMetric.Values(
-              metrics.hasCount() ? metrics.getCount() : null,
-              metrics.hasMinimum() ? metrics.getMinimum() : null,
-              metrics.hasMaximum() ? metrics.getMaximum() : null,
-              metrics.hasMean() ? metrics.getMean() : null,
-              metrics.hasMedian() ? metrics.getMedian() : null,
-              metrics.hasMode() ? metrics.getMode() : null,
-              metrics.hasSum() ? metrics.getSum() : null);
-        } else {
-          assert false : "branch not covered";
-        }
-        if (value != null) {
-          properties.put(property, value);
-        }
-      }
-    }
-    var result = new AggregateResponse(properties, totalCount);
-    return result;
+    return new AggregateUnmarshaler(reply).single();
   }
 
   public AggregateGroupByResponse nearVector(
       Float[] vector,
-      Consumer<NearVector.Builder> nearVectorOptions,
-      Consumer<AggregateGroupByRequest.GroupBy.Builder> groupByOptions,
+      AggregateGroupByRequest.GroupBy groupBy,
       Consumer<AggregateRequest.Builder> options) {
     var aggregation = AggregateRequest.with(collectionName, options);
-    var nearVector = NearVector.with(vector, nearVectorOptions);
-    var groupBy = AggregateGroupByRequest.GroupBy.with(groupByOptions);
+    var nearVector = NearVector.with(vector, opt -> {
+    });
 
     var req = new AggregateMarshaler(aggregation.collectionName())
         .addAggregation(aggregation)
@@ -176,49 +67,23 @@ public class WeaviateAggregate {
         .addNearVector(nearVector)
         .marshal();
     var reply = grpcClient.grpc.aggregate(req);
+    return new AggregateUnmarshaler(reply).grouped();
+  }
 
-    List<Group<?>> groups = new ArrayList<>();
-    if (reply.hasGroupedResults()) {
-      for (var result : reply.getGroupedResults().getGroupsList()) {
-        final Long totalCount = result.hasObjectsCount() ? result.getObjectsCount() : null;
+  public AggregateGroupByResponse nearVector(
+      Float[] vector,
+      Consumer<NearVector.Builder> nearVectorOptions,
+      AggregateGroupByRequest.GroupBy groupBy,
+      Consumer<AggregateRequest.Builder> options) {
+    var aggregation = AggregateRequest.with(collectionName, options);
+    var nearVector = NearVector.with(vector, nearVectorOptions);
 
-        GroupedBy<?> groupedBy = null;
-        var gb = result.getGroupedBy();
-        if (gb.hasInt()) {
-          groupedBy = new GroupedBy<Long>(gb.getPathList().get(0), gb.getInt());
-        } else if (gb.hasText()) {
-          groupedBy = new GroupedBy<String>(gb.getPathList().get(0), gb.getText());
-        } else {
-          assert false : "branch not covered";
-        }
-
-        Map<String, Metric.Values> properties = new HashMap<>();
-        for (var agg : result.getAggregations().getAggregationsList()) {
-          var property = agg.getProperty();
-          Metric.Values value = null;
-
-          if (agg.hasInt()) {
-            var metrics = agg.getInt();
-            value = new IntegerMetric.Values(
-                metrics.hasCount() ? metrics.getCount() : null,
-                metrics.hasMinimum() ? metrics.getMinimum() : null,
-                metrics.hasMaximum() ? metrics.getMaximum() : null,
-                metrics.hasMean() ? metrics.getMean() : null,
-                metrics.hasMedian() ? metrics.getMedian() : null,
-                metrics.hasMode() ? metrics.getMode() : null,
-                metrics.hasSum() ? metrics.getSum() : null);
-          } else {
-            assert false : "branch not covered";
-          }
-          if (value != null) {
-            properties.put(property, value);
-          }
-        }
-        Group<?> group = new Group<>(groupedBy, properties, totalCount);
-        groups.add(group);
-
-      }
-    }
-    return new AggregateGroupByResponse(groups);
+    var req = new AggregateMarshaler(aggregation.collectionName())
+        .addAggregation(aggregation)
+        .addGroupBy(groupBy)
+        .addNearVector(nearVector)
+        .marshal();
+    var reply = grpcClient.grpc.aggregate(req);
+    return new AggregateUnmarshaler(reply).grouped();
   }
 }
